requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.31.0
gunicorn==21.2.0
Pillow==10.0.1
whitenoise==6.5.0
